QtQuick_OpenNI_NiTE_Sample1 rendering improved

+++ QtQuick_OpenNI_NiTE_Sample1/AbstractScene.h
@@ -7,7 +7,7 @@ public:
     virtual void initialise() = 0;
     virtual void update(float t) = 0;
     virtual void render() = 0;
-    virtual void resize( int w, int h ) = 0;
+    virtual void resize( float w, float h ) = 0;
 };
 
 #endif // ABSTRACTSCENE_H
